Test package conversion with shared object files
Right now the test suite consistently fails (or rather, is expected to
fail consistently :-) on Travis CI as described by me earlier today in
issue #3. However while researching I noticed that build 3 on Travis CI
actually passed for Python 3.4 which it really shouldn't have
according to my expectations!

After some research I realized the "problem". The test suite doesn't
explicitly include a test with a Python package containing shared
object files, it only includes such a package implicitly:

 * test_conversion_of_extras converts raven[flask]==3.6.0
 * raven[flask]==3.6.0 pulls in Flask
 * Flask pulls in markupsafe
 * markupsafe includes markupsafe/_speedups.so,
   BUT ONLY ON PYTHON 2.X!

I've now rectified this situation by explicitly including the conversion
of a Python package that is know to require a compiled shared object
file. That package (setproctitle) supports Python 3.x as well so this
test can be run on all supported Python versions.

The immediate effect is that this will break the one passing test suite
version on Travis CI :-) but that is exactly my intention: I'd rather
have a failing test suite than a false sense of safety.

For the record: The complete test suite passes for me locally on all
documented to be supported Python versions, but for more details about
that please refer to issue #3.

  $ tox
  ...
  py26: commands succeeded
  py27: commands succeeded
  py34: commands succeeded
  congratulations :)

Issue 3 on GitHub:
  https://github.com/paylogic/py2deb/issues/3

Build 3 on Travis CI:
  https://travis-ci.org/paylogic/py2deb/builds/52351852