See what LB_LIBRARY_PATH should be set to on Travis CI
Related to issue #3 on GitHub:
  https://github.com/paylogic/py2deb/issues/3

The test suite is now passing on Python 2.7 and 3.4 but its still failing
on Python 2.6. I'm not sure why the difference, but it still seems to fail
on the exact same issue. The simplest workaround seems to be to set
LB_LIBRARY_PATH for Travis CI runs. But then I need to know what I should
set it to :-).

See also the following StackOverflow thread for more details:
  http://stackoverflow.com/questions/23244418/set-ld-library-path-before-importing-in-python