Switch from cached-property to property-manager
The py2deb project comes from a time (2013) when Python descriptors and
cached properties were still magic to me, however since then I created
the property-manager project (2015).

At this point in time (2018) several of the dependencies of py2deb
(other projects of mine) already use property-manager and the
integration of property-manager in py2deb can help to improve the
project, so this seemed like the logical choice.