Improve dependency installation on Travis CI
Even though build 713285687 [1] succeeded I noticed an inconsistency
that slows down builds (in the best case) and may actually cause
spurious failures (in the worst case):

  $ export PYTHON=python3.5
  ...
  $ pip install -r requirements-travis.txt
  DEPRECATION: Python 2.7 reached the end of its life on January 1st, 2020.

This happened because the 'pip install' command was listed in the Travis
CI build configuration (.travis.yml) which isn't aware of the virtual
environment managed by the make file. By coincidence a 'pip' executable
does exist in the system wide search path and so that is used,
completely bypassing the virtual environment... 😬

By moving the 'pip install' command to the makefile, which is aware of
the virtual environment it is managing, I'm hoping to avoid this problem.

[1] https://travis-ci.org/github/paylogic/py2deb/jobs/713285687